requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
arrow==1.2.3
    # via gitlint-core
asgiref==3.7.2
    # via django
attrs==23.1.0
    # via flake8-bugbear
click==8.1.3
    # via gitlint-core
django==4.2.7
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-filter
    #   django-ninja
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==23.3
    # via -r requirements.in
django-ninja==1.0.1
    # via -r requirements.in
factory-boy==3.3.0
    # via -r requirements.in
faker==20.1.0
    # via factory-boy
flake8==6.1.0
    # via
    #   -r requirements.in
    #   flake8-broken-line
    #   flake8-bugbear
    #   flake8-commas
    #   flake8-docstrings
    #   flake8-modern-annotations
    #   flake8-pyproject
    #   flake8-quotes
flake8-broken-line==1.0.0
    # via -r requirements.in
flake8-bugbear==23.9.16
    # via -r requirements.in
flake8-commas==2.1.0
    # via -r requirements.in
flake8-docstrings==1.7.0
    # via -r requirements.in
flake8-modern-annotations==1.5.0
    # via -r requirements.in
flake8-pyproject==1.2.3
    # via -r requirements.in
flake8-quotes==3.3.2
    # via -r requirements.in
gitlint==0.19.1
    # via -r requirements.in
gitlint-core[trusted-deps]==0.19.1
    # via
    #   gitlint
    #   gitlint-core
isort==5.12.0
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy==1.7.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
psycopg[binary]==3.1.13
    # via -r requirements.in
psycopg-binary==3.1.13
    # via psycopg
pycodestyle==2.11.1
    # via flake8
pydantic==2.5.1
    # via django-ninja
pydantic-core==2.14.3
    # via pydantic
pydocstyle==6.3.0
    # via flake8-docstrings
pyflakes==3.1.0
    # via flake8
pyjwt==2.8.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   faker
sh==1.14.3
    # via gitlint-core
six==1.16.0
    # via python-dateutil
snowballstemmer==2.2.0
    # via pydocstyle
sqlparse==0.4.4
    # via django
typing-extensions==4.8.0
    # via
    #   flake8-modern-annotations
    #   mypy
    #   psycopg
    #   pydantic
    #   pydantic-core